[SD-253] Move ERwComp to Snowdrop.Mempool
