+++ snowdrop-core/snowdrop-core.cabal
@@ -69,7 +69,6 @@ library
       Snowdrop.Core.ERoComp
       Snowdrop.Core.ERoComp.Helpers
       Snowdrop.Core.ERoComp.Types
-      Snowdrop.Core.ERwComp
       Snowdrop.Core.Expand
       Snowdrop.Core.Expand.Type
       Snowdrop.Core.Expand.Sequential
