+++ snowdrop-mempool/src/Snowdrop/Mempool/ERwComp/IOExecutor.hs
@@ -1,0 +1,34 @@
+module Snowdrop.Mempool.ERwComp.IOExecutor
+       ( runERwCompIO
+       ) where
+
+import           Universum
+
+import           Data.Default (Default)
+
+import           Loot.Base.HasLens (HasLens')
+
+import           Snowdrop.Core (BException, ChgAccum, Ctx, HasBException, StatePException)
+import           Snowdrop.Dba.Base (DbActions (..), IOCtx (..), runERoCompIO)
+import           Snowdrop.Mempool.ERwComp.Type (ERwComp, runERwComp)
+import           Snowdrop.Util (ExecM)
+import qualified Snowdrop.Util as Log
+
+runERwCompIO
+  :: forall s conf da daa ctx m a.
+    ( Show (BException conf)
+    , Typeable (BException conf)
+    , Default (ChgAccum conf)
+    , HasBException conf StatePException
+    , MonadIO m
+    , MonadReader ctx m
+    , HasLens' ctx Log.LoggingIO
+    , DbActions da daa (ChgAccum conf) ExecM
+    , Ctx conf ~ IOCtx da (ChgAccum conf)
+    )
+    => daa ExecM
+    -> s
+    -> ERwComp conf da s a
+    -> m (a, s)
+runERwCompIO daa initS comp = runERoCompIO daa Nothing (runERwComp comp initS)
+
